requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
python-multipart==0.0.9
langchain==0.2.5
langchain-community==0.2.5
langchain-groq==0.1.5
langchain-huggingface==0.0.3
langchain-pinecone==0.1.3
sentence-transformers==3.0.1
pymupdf==1.24.5
tiktoken==0.7.0
streamlit==1.35.0
pinecone-client